Z_MallocStaticWithUser i32 -> u16

+++ z_zone.c
@@ -544,7 +544,7 @@ void __far* Z_MallocStatic(uint16_t size)
 }
 
 
-void __far* Z_MallocStaticWithUser(int32_t size, void __far*__far* user)
+void __far* Z_MallocStaticWithUser(uint16_t size, void __far*__far* user)
 {
 	return Z_Malloc(size, PU_STATIC, user);
 }
